+++ src/mmds/src/persist.rs
@@ -14,6 +14,7 @@ use super::ns::MmdsNetworkStack;
 
 /// State of a MmdsNetworkStack.
 #[derive(Clone, Versionize)]
+// NOTICE: Any changes to this structure require a snapshot version bump.
 pub struct MmdsNetworkStackState {
     mac_addr: [u8; MAC_ADDR_LEN],
     ipv4_addr: u32,